services/persistence_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
openpyxl
pandas
python-multipart
pydantic